feat: 重构组件，移除设计系统，新增独立组件和常量文件
