+++ app/index.tsx
@@ -1,3 +1,5 @@
+import { Badge } from "@/components/Badge";
+import { Card } from "@/components/Card";
 import {
   AuxChain,
   ChainData,
@@ -5,15 +7,11 @@ import {
   PrecedentCase,
   PrecedentManager,
 } from "@/components/CTDPComponents";
-import {
-  Badge,
-  Card,
-  Colors,
-  FontSizes,
-  ProgressBar,
-  Spacing,
-} from "@/components/DesignSystem";
+import { ProgressBar } from "@/components/ProgressBar";
 import { Pattern, SteadyStateAnalysis } from "@/components/RSIPComponents";
+import { Colors } from "@/constants/colors";
+import { FontSizes } from "@/constants/fontSize";
+import { Spacing } from "@/constants/spacing";
 import { Ionicons } from "@expo/vector-icons";
 import React, { useState } from "react";
 import {
@@ -68,7 +66,7 @@ export default function Dashboard() {
     },
   ]);
 
-  const [patterns, setPatterns] = useState<Pattern[]>([
+  const [patterns, _setPatterns] = useState<Pattern[]>([
     {
       id: "pattern-1",
       title: "回家立即洗澡",
