+++ constants/borderRadius.ts
@@ -1,0 +1,8 @@
+export const BorderRadius = {
+  sm: 4,
+  md: 8,
+  lg: 12,
+  xl: 16,
+  "2xl": 24,
+  full: 9999,
+};
