+++ constants/colors.ts
@@ -1,0 +1,36 @@
+export const Colors = {
+  primary: "#6366f1",
+  primaryLight: "#a5b4fc",
+  primaryDark: "#4338ca",
+  secondary: "#10b981",
+  secondaryLight: "#6ee7b7",
+  danger: "#ef4444",
+  dangerLight: "#fca5a5",
+  warning: "#f59e0b",
+  warningLight: "#fcd34d",
+  success: "#10b981",
+  successLight: "#6ee7b7",
+
+  // 中性色
+  gray50: "#f9fafb",
+  gray100: "#f3f4f6",
+  gray200: "#e5e7eb",
+  gray300: "#d1d5db",
+  gray400: "#9ca3af",
+  gray500: "#6b7280",
+  gray600: "#4b5563",
+  gray700: "#374151",
+  gray800: "#1f2937",
+  gray900: "#111827",
+
+  // 背景色
+  background: "#ffffff",
+  backgroundSecondary: "#f9fafb",
+  backgroundTertiary: "#f3f4f6",
+
+  // 文本色
+  textPrimary: "#111827",
+  textSecondary: "#6b7280",
+  textTertiary: "#9ca3af",
+  textInverse: "#ffffff",
+};
